moved outpoint to a separate file
